pyproject.toml
[project]
name = "zh-context-scanner"
version = "1.0.0"
description = "CLI/TUI tool for detecting and translating Chinese text in source code"
requires-python = ">=3.10"
dependencies = [
    "rich>=15.0.0",
    "regex>=2026.4.4",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "openai>=2.32.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
zh-scanner = "src.main:entry"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]